move components into feature folder to set up for state management
